requirements.txt
flask
google-cloud-aiplatform
numpy
scikit-learn